THE VERSION TREE
In this assignment, I worked with tree structures. The case study that I consider is a generic version tree
that can be used to record different versions of software. Software configuration management (SCM) is the process of
tracking and controlling changes in the software. The version control system can be defined as software that allows
you to create versions of elements and be involved in different versions independently from each other. The overall
process of introducing versions involves both their creation and the implementation of structures for their storage.
Typically, the structures used are either chains or trees.